requirements.txt
openai>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
asyncio
numpy>=1.24.0
scipy>=1.10.0
pandas-ta>=0.3.14
ta>=0.11.0
scikit-learn>=1.3.0
TA-Lib>=0.4.24
hyperliquid-python-sdk>=0.21.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0